feat: add ticket verification, QR code email embedding, and inventory management
Implement ticket verification endpoint and enhance ticket generation with
QR codes embedded in emails via CID. Add inventory tracking and sold-out
handling for ticket types.

Event Service Changes:
- Add ticket verification endpoint (GET /api/v1/tickets/verify/{ticketCode})
- Create TicketService and TicketServiceImpl for ticket verification logic
- Add TicketVerificationResponse DTO for verification result
- Configure SqsClient bean in AWSConfiguration for message publishing
- Add JacksonConfiguration to handle Java 8 date/time serialization
- Update SecurityConfig to permit public access to ticket verification endpoint

Ticket Generation & QR Code Enhancements:
- Generate QR codes containing verification URL instead of ticket code
- Change QR code to encode ALB base URL + ticket verification endpoint
- Add application.alb.url configuration property for dynamic URL generation

Inventory Management:
- Add ticket availability validation before registration
- Track sold ticket count and update TicketType.soldCount after registration
- Auto-disable ticket types when sold out (soldCount equals quantity)
- Change TicketType quantity and soldCount from Integer to Long
- Add remainingTickets field to TicketTypeResponse DTO
- Update TicketTypeMapper to calculate remaining tickets

Email Notification Improvements:
- Implement CID (Content-ID) embedded images for QR codes in emails
- Add sendEmailWithQRCodes method to embed QR codes as inline attachments
- Extract base64 data from QR code URLs and attach as ByteArrayDataSource

Configuration Updates:
- Add sqs.payment-processing-event-queue-url property
- Add sqs.payment-completed-event-queue-url property
- Rename TICKET_PURCHASED_EVENT_QUEUE_NAME to TICKET_PURCHASED_EVENT_QUEUE_URL
- Add application.alb.url for ticket verification endpoint generation
- Configure JavaTimeModule for proper Instant serialization in JSON

Technical Improvements:
- Add comprehensive JavaDoc comments for service methods
- Add detailed logging for SQS message publishing
- Use injected ObjectMapper with JavaTimeModule for date handling
- Improve null safety with Boolean.FALSE.equals() checks

+++ services/event-service/src/main/java/com/event_service/event_service/models/TicketType.java
@@ -29,9 +29,9 @@ public class TicketType {
 
     private Double price;
 
-    private Integer quantity;
+    private Long quantity;
 
-    private Integer soldCount;
+    private Long soldCount;
 
     private Boolean isActive;
